+++ console/src/layout/components/TagsView/ScrollPane.vue
@@ -73,7 +73,7 @@ export default {
   position: relative;
   overflow: hidden;
   width: 100%;
-  /deep/ {
+  :deep() {
     .el-scrollbar__bar {
       bottom: 0px;
     }