+++ console/src/store/modules/tagsView.js
@@ -9,7 +9,7 @@ const mutations = {
     state.visitedViews.push(
       Object.assign({}, view, {
         title: view.meta.title || 'no-name',
-      })
+      }),
     )
   },
   ADD_CACHED_VIEW: (state, view) => {