+++ console/src/store/modules/user.js
@@ -1,4 +1,4 @@
-import { login, getInfo, getLoginOptions } from '@/api/user'
+import { login, logout, getInfo, getLoginOptions } from '@/api/user'
 import { getToken, setToken, removeToken } from '@/utils/auth'
 import { resetRouter } from '@/router'
 const _ = require('lodash')
@@ -109,15 +109,15 @@ const actions = {
   // user logout
   logout({ commit, state }) {
     return new Promise((resolve, reject) => {
-      // logout(state.token).then(() => {
-      commit('SET_TOKEN', '')
-      // commit('SET_ROLES', [])
-      removeToken()
-      resetRouter()
-      resolve()
-      // }).catch(error => {
-      //   reject(error)
-      // })
+      logout(state.token).then(() => {
+        commit('SET_TOKEN', '')
+        // commit('SET_ROLES', [])
+        removeToken()
+        resetRouter()
+        resolve()
+      }).catch(error => {
+        reject(error)
+      })
     })
   },