+++ server/conf/config.js
@@ -10,6 +10,9 @@ const config = {
   database: {
     url: process.env.RBAC_SQL_URL || 'postgres://wolfroot:123456@127.0.0.1:5432/wolf',
   },
+  redis: {
+    url: process.env.RBAC_REDIS_URL || 'redis://127.0.0.1:6379/0',
+  },
   oauthOptions: {
     allowEmptyState: false,
     allowExtendedTokenAttributes: true,