+++ server/src/util/token-util.js
@@ -2,7 +2,7 @@
 const config = require('../../conf/config')
 const jwt = require('jsonwebtoken')
 const log4js = require('./log4js')
-
+const {redisClient} = require('./redis-util')
 const ERR_TOKEN_INVALID = 'ERR_TOKEN_INVALID'
 
 const tokenVersion = 2
@@ -39,5 +39,39 @@ function tokenDecrypt(token) {
   }
 }
 
+async function tokenCreate(userInfo, appid) {
+  const tokenInfo = tokenEncrypt(userInfo, appid)
+  const token = tokenInfo.token
+  const expiresIn = tokenInfo.expiresIn
+  // Because the key already contains all the information.
+  const res = await redisClient.set(`wtk:${token}`, '1', {'EX': expiresIn});
+  if (res !== 'OK') {
+    throw new Error('redis set error');
+  }
+  return tokenInfo
+}
+
+async function tokenCheck(token) {
+  const user = tokenDecrypt(token)
+  if(user.error) {
+    return user
+  }
+  const res = await redisClient.get(`wtk:${token}`);
+  if (res !== '1') {
+    return { error: ERR_TOKEN_INVALID }
+  }
+  return user
+}
+
+async function tokenDelete(token) {
+  const res = await redisClient.del(`wtk:${token}`);
+  if (res !== 1) {
+    throw new Error('redis delete error');
+  }
+}
+
 exports.tokenEncrypt = tokenEncrypt
 exports.tokenDecrypt = tokenDecrypt
+exports.tokenCreate = tokenCreate
+exports.tokenCheck = tokenCheck
+exports.tokenDelete = tokenDelete